Modificado maneira de salvar a imagem ao adicionar um novo jogo. Aplicado melhoria ao retornar a imagem na rota de buscar um jogo. Agora a rota de remover um jogo também remove a imagem.

+++ helpers.py
@@ -1,13 +1,18 @@
 import os
 from main import app
 
+
 def recupera_imagem(id):
     for nome_arquivo in os.listdir(app.config['UPLOAD_PATH']):
-        if f'capa{id}' in nome_arquivo:
+        boolean = f'{id}.jpg' == nome_arquivo
+        if boolean:
             return nome_arquivo
 
-    return 'capa_padrao.jpg'
+
 
 def deleta_arquivo(id):
     arquivo = recupera_imagem(id)
-    os.remove(os.path.join(app.config['UPLOAD_PATH'], arquivo))
+    if arquivo is None:
+        return
+    else:
+        os.remove(os.path.join(app.config['UPLOAD_PATH'], arquivo))
